requirements.txt
Django == 3.2.16
django_multiselectfield == 0.1.12
django-allauth == 0.43.0
Pillow == 7.2.0
Unidecode == 1.1.2
urllib3 == 1.25.10
requests == 2.27.1
dj-static == 0.0.6
pymysql == 1.0.2